+++ common/tests/fakes.py
@@ -90,7 +90,8 @@ class MatchRequestFactory(DjangoModelFactory):
     topic_channel = factory.SubFactory(TopicChannelFactory)
     profile = factory.SubFactory(SocialProfileFactory)
 
-
+# NOTE: this is the best example of how to fake out models that refer to other models the SubFactory corresponds to a
+# foreign key field and the @factory.post_generation corresponds
 class MatchFactory(DjangoModelFactory):
     class Meta:
         model = Match
