+++ matchmaking/management/commands/remind_matches.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
-from matchmaking.tasks import remind_matches
+from matchmaking.common import remind_matches
 
 class Command(BaseCommand):
     help = """Remind matches without Penny Chats to set one up."""
