--- matchmaking/tests/tasks.py
@@ -1,4 +1,0 @@
-
-
-
-#TODO! test request_matches
